Avoid string copy in set_mode
Summary:
`bumpalo::collections::String::from_utf8_lossy_in` is different from `std::string::String::from_utf8_lossy`. The std version returns a `Cow<str>`--it is either a reference to the input if the input is valid UTF-8, or an owned `String` otherwise. The bumpalo version unconditionally copies the input into the arena.

We don't actually need to keep the contents of the string--we're only looking for the substrings "decl", "partial", or "strict". If we wanted to hold on to the contents (including when they have invalid UTF-8 characters), from_utf8_lossy_in might be a good choice. But since we don't, `std::str::from_utf8` (which never makes a copy) is sufficient.

Strictly speaking, we don't need to check for valid UTF-8 here at all--we could look for the mode substring in a byte string instead. But it is convenient to have access to the `trim_start_matches` and `trim` methods on `&str`, so I have not tried to get rid of the `str`-conversion.

Reviewed By: Wilfred

Differential Revision: D21645323

fbshipit-source-id: c8c12f96701a7988a4ffe209e58ff72277228992