Panic instead of returning Err
Summary:
The decl parser should never fail--no matter how nonsensical the input file, we should return our best possible attempt at representing a decl for it.

Panicking is not the right strategy for most of these cases (the FFP might pass us some of these unexpected node kinds in files with parse errors), but it's useful as a way to surface them. In future diffs, I'll replace many of these panics with returning Node_::Ignored.

Reviewed By: Wilfred

Differential Revision: D21693157

fbshipit-source-id: dae471570951c0ffb11692254b59b5d026ed2372