Reduce size of Node from 56 bytes to 32
Summary: Put the fields of Node_ variants with three-, four-, and five-qword fields behind a pointer, to reduce the size of every variant of Node_ from 48 bytes (six qwords--one for the tag, and five for the largest variant's fields) to 24 (three qwords--one for the tag, and two for the fields). The Node type (an alias for Result<Node_, String>) is one qword larger (for the `Ok` tag).

Reviewed By: Wilfred

Differential Revision: D21688547

fbshipit-source-id: e24429a02ed60c33f97213fd0afce2bc5a4f0b82