Implement NodeType for Node
Summary:
In some circumstances, the parser needs to examine a recently-parsed node in order to determine what kind of node it is. This informs how the parser will proceed (e.g., whether it will choose to backtrack from a speculative parse or not).

Since the parser is generic over its smart constructors, and thus generic over the type of node they produce, the NodeType trait specifies the interface that nodes must satisfy, enabling the parser to inspect them.

Not all implementors of the SmartConstructors trait are interested in maintaining knowledge of what kind each node is. So for convenience, [an implementation of NodeType is provided for `(SyntaxKind, T)`](https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/hphp/hack/src/parser/smart_constructors.rs?commit=b12f046b1ef975dbc7fa90424cf951b42c63d297&lines=39), and a [SmartConstructors implementation](https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/hphp/hack/src/parser/smart_constructors_wrappers.rs?commit=b12f046b1ef9&lines=1) using that node type is provided as an adapter. This allows the facts parser's smart constructors, which ignore most nodes, to be used by the parser--the knowledge of what kind a node was is maintained by the wrapper, even if the facts smart constructors chose to ignore it.

One disadvantage of the wrapper is that it inflates the size of the nodes used by the parser. If type T has alignment 8, then (SyntaxKind, T) will be 8 bytes larger than T. This is the case in the direct decl parser--while Node is 16 bytes, the values handled by the parser are of type (SyntaxKind, Node), which is 24 bytes. This costs performance, but perhaps more importantly, it costs stack space, which is at a premium in the FFP. Nodes are stack-allocated until they can be passed to the constructor of their parent node, and the parser recurses very deeply on some files. Reducing the size of nodes means we need to worry less about running out of stack space.

We can reduce the size of these values by doing away with the WithKind wrapper, but to do so, we need to track the knowledge of whether a node was one of these kinds ourselves, and expose it by implementing NodeType for Node. This diff does so, by adding an IgnoredSyntaxKind variant for nodes where we need only track the kind.

Because we now represent ignored nodes in multiple ways (either Ignored or IgnoredSyntaxKind(..)), this diff also updates every place where we matched against Ignored to invoke is_ignored (or the new helper is_present) instead.

Reviewed By: Wilfred

Differential Revision: D21696085

fbshipit-source-id: ae125c3b3eba8c1804864b53e78985e2e2009793