Fix two bugs in keyset prepend
Summary:
SetArray::Prepend doesn't behave correctly. It allows us to put multiple copies of the same key into a keyset, and it incorrectly advances the keyset's internal cursor (all other array-likes reset their cursor on prepend). These bugs are hard to spot because we implement prepend in a different way in ext_array for non-vector-data keysets. As a result, they are probably incredibly rare in practice - who wants a keyset with the keys 0...(n-1) in order? But we still need to fix them.

The only real subtlety in the fix is what we do about key ordering if we prepend an existing key. I've made the code for vector-data keysets do the same thing we do for non-vector-data keysets: the newly prepended element is removed and re-inserted at the beginning of the keyset.

I detected the bugs because of assertions in LoggingArray - essentially, the bespoke virtualization system is acting like a fuzzer on all sorts of array ops.

Reviewed By: ricklavoie

Differential Revision: D21873922

fbshipit-source-id: de91d4a1c0972b407a2329b8e61ba53519d3c79d