Fix a couple one-off bespoke issues
Summary:
For the varray and darray conversion results, we can only make an assertion about layout for vanilla arrays.

There's an optimization to skip ksort for PackedArray. This optimization should be based on type (varray and vec), not on layout; it applies equally well to bespoke varrays and vecs. (All PackedArrays are varrays or vecs, so it's a strict increase in coverage.) We had to weaken some assertions in ext_array due to this optimization; with the alternate type-level version of the optimization, we can strengthen them.

Reviewed By: ricklavoie

Differential Revision: D21872166

fbshipit-source-id: 5f75ffa2848a057071c40c5c752d8e9cafeb3233