Handle bespokes in mutable member ops
Summary:
Add support for mutable member ops on bespoke array-likes. There are five big implementation blocks here: ElemD, ElemU, SetElem, SetNewElem, and UnsetElem. IncDecElem and SetOpElem can reuse ElemD.

There main question for this diff is where to put the bespoke branch: at the beginning of the op, or in each DataType handler? In general, I have a preference for doing it at the beginning, or right after the array case if there is one (since the array case is already virtual, it usually handles bespokes correctly). For SetElem, I opted to put it in the vec and dict cases instead.

There are a few cases where the ArrayKind helper throws a different error than the member-operations logic. In these cases, I've changed the helper to match the member-operations, since without this consistency, it's very hard to implement the correct behavior via indirection. This change has no effect on Hack code. It slightly changes the error message (dropping the operation name) for extensions that call .lval() on a keyset - but that's a fundamentally silly thing to do, so I don't care.

This diff causes failures in a RecordArray test: because of the way I've arranged ArrayKinds, a RecordArray actually counts as !isVanilla (and then proceeds to crash all over). I don't plan to fix this failure; we're deprecating RecordArray instead. I'm deleting the test for now. Confirmed with arnabde03 that that's okay.

Reviewed By: ricklavoie

Differential Revision: D21868300

fbshipit-source-id: aed592452fee519799f5db27ad9ed31ef9303077