Fix bug in get_frame_args
Summary:
We should never use array layouts for logic, only for optimizations. This line also isn't HAM-safe - post-HackArrDVArrs, the varargs will be in a vec.

Use a type predicate for varargs instead of a layout predicate. "isVecOrVArray" uses the HAM flag flip to determine which type to check - it doesn't blindly pass both vecs and varrays.

This code is only used for slow interception / introspection cases, so perf is trivial.

Reviewed By: ricklavoie

Differential Revision: D21900148

fbshipit-source-id: 88dc2e67c67b7d54169a50ffb6acf49fe9086fa1