Expose RepoAuthType for bespokes
Summary:
RepoAuthType is a claim about the Hack types of data, not about HHVM layouts. As such, it applies to bespokes, too.

Now that we've made this change, there's no reason to set the vanilla bit for the RepoAuthType constructor. It doesn't get used with the AllowBespokeArrayLikes flag off, and we only set it so that we'd actually use the type in this case. We also move the RAT out of the "{...|Bespoke}" union in toString, since it applies regardless of kind status.

Reviewed By: ricklavoie

Differential Revision: D21914104

fbshipit-source-id: ab16a5b5354cc459ce151941233ed2ad63a8d123