HAM-safe reified generics deserialization
Summary:
Something truly horrible has happened in the interaction between HAM and reified generics. Type structure code wants to place tight invariants on the type of its data - in particular, a valid type structure is made of nested dvarrays with the right fields and types. However, when we serialize dvarrays, varrays become darrays...

Deal with this issue in deserialization by changing the type structure code that previously validated arrays to instead validate and coerce. The only coercion we allow is the one that we explicitly broke by deserializing these arrays: the vector-like darray to varray coercion.

In the process, clean up this code. The key cleanup is in coerceTypeStructure (formerly: isValidResolvedTypeStructure), which now reads somewhat declaratively as a simple list of constraints on the given fields. We do that by pulling the actual coercion logic into helpers so that all we need to say in the main function is "this field is a type structure, and this field is an optional list of type structures".

Reviewed By: ricklavoie

Differential Revision: D21924254

fbshipit-source-id: 8b1d3fa653277f8bee7fd14545bf9102e4e38a18