Fix hole in internal checks
Summary:
In order to make "the big switch" safe, we need to convert internal varray type-checks to vec type checks and internal darray type-checks to dict type-checks. Unfortunately, we have DataType-only "tvIsVecOrArray" and "tvIsDictOrArray" type-checks in our codebase, primarily in type-structure code.

Most of these checks are assertions, and most of them appear in type-structure code, where there's a big type-safety hole addressed in the previous diff. All of the checks should now pass once strengthened - that is, the old and new behavior here (even for the non-assertions) should be the same. Here's a list of non-assertions where it's used:
- In bytecode.cpp, event-hook.cpp, and ext_std_function.cpp: to check the var-args argument. This argument is always a varray.
- In type-struction.cpp and type-structure-helpers.cpp: to check various type-structure fields. Since we fixed the deserialization hole, these fields are the right type.
- In func-emitter.cpp and class.cpp: the reified-generics class attr. This attr uses internal serialization which (I believe) preserves dvarrays.
- To handle two reified-generics bytecodes: these bytecodes always operate on valid TypeStructure lists.

Reviewed By: ricklavoie

Differential Revision: D21923281

fbshipit-source-id: e25248d05bbaa1e5a550825b990ab1417d00a26c