Rename `*Like` types to `*Compat`
Summary: Choosing the names `ArrLike` and `VecLike` for the `T | ClsMeth` types is ... really problematic. As per kshaunak on D21596387 there's already internal ambiguity where `ArrLike` sometimes means "the CoW arrays." This just renames the `ArrLike` types to `ArrCompat`. This is going to be the base for D21596387 where instead of renaming a bunch of the builtin names we just formalize them and add them to the `trep` enum.

Reviewed By: viratyosin

Differential Revision: D21830917

fbshipit-source-id: c55ab8223380a9cb24008d3622f13998df606cef