Move `is_any_array` types into `trep` (+ rename)
Summary: For some upcoming changes I'm making I'd like to align how hhbbc treats `IsTypeOp::Arr` and `IsTypeOp::ArrLike`. The best way to do that is to formalize the "any array" type in HHBBC.  Unfortunately there's a clash where `ArrLike` has come to mean "an array or a clsmeth." I've renamed the new `trep` entries to `AnyArr*`, and updated HHVM to use the new names and use the new types when reasoning about `IsTypeOp::ArrLike`.

Differential Revision: D21596387

fbshipit-source-id: 3dda729b954ef371aea1839878335479b15a05b0