Introduce StFrameCtx
Summary: Add an instruction specifically for writing to the `m_thisUnsafe` field of the ActRec. The final diff in this stack eliminates `DefInlineFP`, rather than write to the `this` location in the same instruction that the frame is instantiated this the diff will use a single PureStore to update `this` to allow store-elim to elide or move the store.

Reviewed By: jano

Differential Revision: D21389088

fbshipit-source-id: 35684eb17382d5fcb6a76ec4b9675b8acadcf80e