Rename WatchmanAutoload namespace to Facts namespace
Summary:
The original namespace for all native autoloading stuff was `WatchmanAutoload`. I chose this name because "native autoloading" is not very descriptive in a codebase where all code is native, and where we expect a totally separate production autoloader (also native) to run. The distinguishing feature of my native autoloader was that it used Watchman to track change files, so that's the name I chose.

Now that I'm sure that this extension is going to be used for more than autoloading, and that this extension won't always be using Watchman to track file reads, I think it's time to rename the namespace to `Facts`. This has an added bonus of helping issue triage, as fewer stack traces will contain the string "Watchman" in them.

Reviewed By: dneiter

Differential Revision: D21943002

fbshipit-source-id: ec67f424d54bd92706cc591c2c246ee2551c62df