Update kind on conversion
Summary:
From the slow bespoke tests, another easy correctness fix: when we perform a conversion op on a LoggingArray and get the same inner array back, we may still have changed its kind (e.g. packed -> vec) and must change the outer array's kind too. This issue doesn't happen for packed to mixed because it's always a new array in that case.

m_kind is protected, so I added a helper method that does the update.

Reviewed By: jano

Differential Revision: D21964056

fbshipit-source-id: 6e0af880a534210da363c0b0bd55a848a4ba5a44