Add eval mode to hhvm binary and use it in php_wrapper to implement -r
Summary: We want to lift top level code from `php -r <code>` into an `__EntryPoint` function so that users don't need to write out the function by hand once top level code is banned. We already do this for the debugger REPL, so add a new mode to hhvm to run a string as a debugger eval string and use it in php_wrapper.

Reviewed By: oulgen, jano

Differential Revision: D21957992

fbshipit-source-id: 222027a987838da3c2c78b6862a86653559fae49