yet more logging for client connect
Summary: I'm trying to track down monitor connection failures. This diff adds a few lines more logging to the client log. A lot of it is just minor tweaks and formatting. The only invasive change is that now, for each attempt we make to connect to the monitor, it associates that attempt with a unique connection-attempt-id in the client log.

Reviewed By: arxanas

Differential Revision: D21875537

fbshipit-source-id: a79504fec4326b19ec7009df39e61e02850acbc8