Disable bespoke when array provenance is used
Summary:
I don't have the heart to make bespoke arrays work with array provenance right now. Instead, we'll disable this optimization when we turn on array provenance logging. Note that we can still do a repo build with provenance on and turn on bespokes at runtime - that's our prod setup.

There are a few tests that examine provenance logs without Eval.LogArrayProvenance on. I'm fine with losing provenance coverage when this flag isn't on. It doesn't get used. To make those tests work, I turned the runtime instrumentation on. (As a follow-up, we may want to make get_provenance return empty when the runtime instrumentation is off.)

Reviewed By: ricklavoie

Differential Revision: D21968704

fbshipit-source-id: 6713e81000bcca3fb0a6305e687c5d64470837f8