Escalate to vanilla in GetScalarArray
Summary:
We use GetScalarArray to explicitly request a static array in two places: first off, when burning specific arrays into the JIT, and second, for reified generics. For both of these cases, it makes more sense to burn an easily-used vanilla array in right now; we can experiment later with static bespokes.

(We do still construct static LoggingArrays as a way to log data about static arrays, but that's done differently...one might say in a bespoke fashion...)

Reviewed By: ricklavoie

Differential Revision: D21968881

fbshipit-source-id: 5c925d2ca9f065a578d6b3016677669cfa90694a