Fix subtle comparison issues
Summary:
PHP comparison is more subtle than one would believe. I originally audited ArrayData::Equal vs. VecEqual, DictEqual, and KeysetEqual, and the same for Lt, found them to be the same, and then dispatched bespokes to the ArrayData method. It turns out that's incorrect for Gt and Gte - they depend on argument order and it's different for PHP arrays vs. vecs! (This could be a problem for HAM, but it's too much of an edge case and unlikely to actually come up.)

There's another issue that I found with bespokes, too. varray[1, NAN] and varray[1, NAN] should compare as not equal because NAN != NAN, but due to static array deduping, they actually compare as equal. There's a test that depends on this behavior, and if static array deduping is tweaked (e.g. with provenance tags, or with bespoke wrappers), it fails. I made that test explicitly depend on identity.

As noted elsewhere, bespokes are a kind of fuzzer on all array behavior...

Reviewed By: ricklavoie

Differential Revision: D21969264

fbshipit-source-id: c2df3041ffbdeb6fcd5137b925c887707e8d2904