Fix error location for inout parameters
Summary:
Add position information to the local variable environment that represents the location that caused the local to get that type, often at an assignment. When checking inout parameters at a return site, this stored position is used as the source of the error.

The only complication is if the return follows a control-flow merge where the type of the inout parameter gets non-trivially merged. At that point it is not clear which branch, if any, will be the problem for the inout, and so we remove the position information, falling back to the location of the return (if any).

Reviewed By: andrewjkennedy

Differential Revision: D21886523

fbshipit-source-id: f6677cfead423c70670fdf3dbbea8ba5fb29c40d