Fix vanilla assumptions in deepCopy
Summary:
IterateKV already optimizes based on packed or mixed layout, so we might as well have a generic deepCopyArrayLike helper. Besides, this deepCopy code is already very cold - I believe it's only used to initialize collections from non-collection objects, or something unusual like that.

If we cared to optimize this code, a much better optimization would be to do an isRefcounted check before calling deepCopyArrayLike. But it's too cold for even that to matter!

Reviewed By: alexeyt

Differential Revision: D21965180

fbshipit-source-id: 21ef68c13ffc5e23464584b96b550ba2b8a956d8