Fix vanilla assumptions in bytecode.cpp
Summary:
iopAddNewElemC assumes that its input is a vanilla vec or keyset. Let's fix that.

While at it, I made this code homogenous in style to iopAddElemC. I also searched for specific calls to PackedArray, MixedArray, and SetArray methods within bytecode.cpp. The only remaining such calls are to constructors, which are safe because they're creating a new vanilla array, not operating on an existing one.

Reviewed By: ricklavoie

Differential Revision: D21964604

fbshipit-source-id: 92e50698994cd2e510edc5424d21cf87b0226e7d