Fix refcounting issue in Map
Summary: I was a fool to try to use a smart pointer here. The behavior of our core array functions is not "homogenous" in refcount - for example, toDict produces a refcount if it creates a new array, but it has no refcounting effect if executes in place. We need to handle each step carefully, and differently.

Reviewed By: alexeyt, leikahing

Differential Revision: D21975816

fbshipit-source-id: b310d118f2d3673d4d1274aecd130ddc3819a7be