Fix vanilla assumptions in SetRange
Summary: This change is a small de-optimization for setRangeVec - specifically, it introduces one extra layout check at the beginning of the loop, before falling into the regular fast PackedArray iteration. It won't matter. setRangeVec and its callers don't appear in prod perf data at all. This code is not completely dead, but it is very nearly so.

Reviewed By: ricklavoie

Differential Revision: D21983476

fbshipit-source-id: f7b7d00764e79657c7614fd6c20646732aadfce2