Output telemetry in JSON format on recheck end
Summary:
I would like to script an end-to-end test that can examine/log the result based on what the telemetry object shows in the output.

This diff adds the recheck telemetry we already have to the recheck stats type that we return to the client.

It's only output when the output format is JSON.

Reviewed By: yinghuitan

Differential Revision: D21724275

fbshipit-source-id: 8c87981e96c7565b0c2004efab9797f33a1fea33