Tiny HAM behavior difference
Summary:
There is a tiny behavior difference between PHP arrays and Hack arrays on inc-dec and set-op final operations: the OutOfBoundsException thrown for PHP arrays includes mention of "Inc/dec" or "Set-op", but the same error for Hack arrays describes the operation as a generic "Lval". IMO we can just drop this difference with no issue. The type of the exception stays the same, and it's not used in ErrorElevator regexes (this exception isn't one we care about for migration purposes).

Why stabilize on the Hack array error? Emitting that error means that we can implement these ops in a simpler way just by calling out to the corresponding ElemD helper.

Why do I care? This error causes a test failure for bespokes, which use the same simple implementation of IncDecElem and SetOpElem that Hack arrays do.

Reviewed By: alexeyt

Differential Revision: D21976925

fbshipit-source-id: 755922833c0f036168877ae9e17bcc4527f4f66b