Add new option stdin-name to specify filename used for stdin input to hh
Summary:
Certain command-line functions to `hh_client` expect the "file to be looked at" in `stdin`. This file is checked alongside the files in the working directory.
A common use is to take a file from the working directory and pipe it to `stdin`. Unfortunately, hh pretends that `stdin` is a file with an empty filename i.e. distinct from any in the working directory. This can lead to duplicate definition errors, which used to be harmless for the purposes of the command-line functions (such as `--identify-function`) but now cause the produced TAST to omit the definitions referenced by the command-line function.

To fix this mess, I introduce a new option `--stdin-name` which lets you specify the "filename" for stdin. This might be the same as one of the files in the working directory, in which case it will take precedence over it (as though you edited it), or it can be distinct. A file corresponding to the filename *must exist* but the contents is ignored.

Reviewed By: vassilmladenov

Differential Revision: D21973403

fbshipit-source-id: ce34aec3052a0969deb123b77cba67935d5e3abc