Fix OOM pre-checking in VariableUnserializer
Summary:
In VariableUnserializer, we attempt to check if deserializing
an array or collection would cause an OOM before doing the
deserialization (this allows us to allocate everything up
front). However the logic had some issues. The biggest problem was
that it was not checking the correct size for packed arrays. Instead
of the size in bytes, it was passing the capacity in elements (which
means it was smaller by a factor of 16). Pass the sizes in bytes
correctly. Second, we were failing to account for the memory allocated
by the m_refs vector. For large arrays this can be significant. Do a
similar pre-OOM check there as well. Finally, we weren't doing any
pre-OOM checking for collections at all. Do that as well.

Differential Revision: D21992350

fbshipit-source-id: 2d9c89a978ff77c068bf40955a25f6b671132957