Fix fbcode/hphp/test/slow/ext_fb/compact_serialize.php when specialization is enforced
Summary:
`fb_(compact_)unserialize` return `darray`s by default. This test was using `===` to compare the original input to the result of round-tripping the input through serialization. When specialization is on, varrays will not equal darrays, causing these equality comparisons with inputs with varrays to fail.

This diff applies a recursive darray cast to the input when comparing to the round-tripped result.

Differential Revision: D22001189

fbshipit-source-id: 51b955ea4ee980297b1435715b3b86fd75dec828