Invert decl fixme code check
Summary:
Adds a whitelist for fixmes that appear outside blocks (referred to as "decl fixmes", but they affect things like top level code as well).

Important conceptual difference between this and {D21608542}: because the semantic position of a fixme isn't propagated to Errors, these fixmes are categorized during parsing and are looked up in different tables.

Reviewed By: Wilfred

Differential Revision: D21668703

fbshipit-source-id: 68fad52abd0a0b426d89c3cdf01c33f4613cf9f4