Fix runtime unit tests
Summary:
loosen_dvarrayness_always is the actual logic of loosening dvarrayness in HHBBC. The unit test should test it, not loosen_dvarrayness. (When we've shipped specialization, we'll rename the "_always" one, dropping the suffix.)

In HHVM unit tests, we need to actually match dvarrayness for equality. We see that:
- varray + varray == darray
- array_merge always returns a plain array
- Array arr; ...; always returns a plain array

We still want to test operations on a varray. However, set-new-key and set-str will both throw, so the only operation we need to test is append.

Reviewed By: ricklavoie

Differential Revision: D22007281

fbshipit-source-id: 759ae90089150e6a88fcde12a506d63dd3eb0b46