Add partial whitelist
Summary:
Same as {D21608542}, but for partial mode. Because a position doesn't carry information about whether a file is partial, this change **breaks <?hh // partial** and relies on the new `.hackpartial` extension to determine whether a file is partial, and therefore which whitelist to look up.

To save unnecessary work, a file extension is only looked up if `allowed_fixme_codes_partial` is set. Otherwise, `allowed_fixme_codes_strict` is the fallback.

This diff has the main restructure of the `add_list` function, and it deletes the code for 5XXX fixmes because Errors doesn't handle those.

Reviewed By: Wilfred

Differential Revision: D21835313

fbshipit-source-id: e5e9e5c873f56c456df54a39d1abf0ebefcfed07