Fix BespokeArray::heapSize issues
Summary:
We can't call allocSize in LoggingArray::MakeFromStatic, because the lad hasn't been initialized yet. However, it's actually more correct to just use size here; for static arrays, we shouldn't do the size class rounding logic. The rounding only applies when we use MemoryManager to allocate memory, but for static arrays, we use an alternate allocator (jemalloc?)

This change clued me in to a couple of other issues. First off, BespokeArray::heapSize just needs to return the actual used size; MemoryManager is already rounding the size up to size class for many kinds (any that don't have a size index stored). Second, allocSize in GetScalarArray is incorrect, because we don't want to round there.

The right fix for GetScalarArray is to have a new virtual dispatch method on arrays to compute the unrounded size. I don't want to do that right now, and it only affects logging, so I left a task instead. Incidentally, the other allocSize method that I deleted in D21854833 rounded PackedArray, but not MixedArray or SetArray, so it was better but still incorrect.

Reviewed By: ricklavoie

Differential Revision: D22014875

fbshipit-source-id: a8be39d2d8ef5a5522485e9ed51820023b5ac33c