Track sizeof(Type) and friends
Summary:
I'm going to land a stack of diffs to cut the size of HHBBC's Type by 3x, from 48 bytes to 16 bytes. My guess is that this will cut HHBBC memory usage by double-digit percentages, since Type is so fundamental. Will try out full repo builds with the stack to confirm.

In this first diff, we simply add static assertions to track the sizes of Type and some of its components.

Reviewed By: ricklavoie

Differential Revision: D22036132

fbshipit-source-id: 5259d41d33f1a2051e6337762c1a4c0ed5e86dcb