Pack Type bits and tag
Summary:
We can cut another quadword out of Type by packing its m_bits field (where only 39 of 64 bits are currently being used) and its data tag byte. This change cuts sizeof(Type) from 40 to 32.

Most of the time when we use m_bits, we're reading it; reading it is type-safe with the new bits() helper. I'm debating whether we should assert that the high bits are unset when writing it; I kept that piece in a separate diff (see the next one in the stack).

Reviewed By: ricklavoie

Differential Revision: D22036144

fbshipit-source-id: 9d2e986998904c419fa5631d4054aff3f11dbd32