Stop writing to the checkout in server tests.
Summary:
Server tests were still writing to the cwd (and therefore potentially the checkout) with the www.pid files that are created in server mode in the CWD by default. Update server_tests.inc to pass a unique -vPidFile value.

For some reason, `./hphp/test/run all` was not running all the server tests, make it do so.

After this diff (modulo T67005367) none of our tests are writing to the checkout.

Reviewed By: ricklavoie

Differential Revision: D22044232

fbshipit-source-id: 048ed797d0a1a78af979bbbfc19565abdbb44bc9