Fix leak in CompactVector
Summary:
Pretty sure this needs fixing...  It's called in rewind in HHBBC, so it could have an effect. The other callers all have types with trivial destructors, though.

I audited the rest of CompactVector and didn't see other issues, but it would be good to have another pair of eyes on it.

Reviewed By: ricklavoie

Differential Revision: D22085622

fbshipit-source-id: 384872a557e72da5e9e5a960be2a550873641c56