align slab allocations
Summary: The first a few bytes in a slab is used for slab metadata. As a result, allocations from slabs are not always aligned to cache line boundaries (even when size of the allocation is a multiple of cache line size). Here we introduce an option to align the effective starting address, which could lead to better cache locality.

Reviewed By: ottoni

Differential Revision: D22059155

fbshipit-source-id: 1640e4b13bf28d863fc7d845085f9e5b9b2fcd75