Support type constraint (typing, part 1)
Summary:
We currently only support reify modifier on case type T statements.
This diff allows to add constraint to such statements.

The ultimate goal is to support things like:
```
class C {
  enum E {
    case type T as num;
    case T val;
    :I(type T = int); // OK
    :S(type S = string); // KO
  }

  public static function add<TP as static:E>(TP $x, TP $y) : num {
    $x = static:E::val($x); // $x:T as num
    $y = static:E::val($y); // $x:T as num
    return $x + $y; // should be num
}
```

In this diff, I introduce the checking of the PU definition (in the example above, checking that `:I` is OK but `:S` is not).
We can't type `C::add` yet. See next diff in the stack.

Reviewed By: andrewjkennedy

Differential Revision: D21879850

fbshipit-source-id: d6c1ad6ab30c0cfad2f2a5b8281af1d229f74414