Fix incorrect custom type-scanning function for UnwindRDS
Summary:
The custom type-scanning function for UnwindRDS was calling scan() on
an expression (the return value of left() or right()). This is, in
general, incorrect, as scan() takes a reference to a pointer, not a
pointer itself. IE, it wants the memory address where the pointer
sits, not the pointer itself. Since scan() is being passed a reference
to a temporary, the address is meaningless, and the GC won't actually
"see" the value. As a result, the GC will free the value if run at the
wrong time.

It's not as simple to just enqueue the pointer instead, as we don't
have an interface for that. The value inside the Either may be masked
so cannot be used directly. For now, just conservative scan it. We
already do this for other problematic cases.

Reviewed By: paulbiss

Differential Revision: D22104885

fbshipit-source-id: a1002ef57e71f21671bec37b9da63717acbdc96f