Add implicit context language feature
Summary:
This diff is the first version of the implicit context language feature.
Design of this feature is available on https://fb.quip.com/2igDArlbyKQ4.
This diff is far from the final version and things on this diff might need to
change as we finalize the design.

This diff does not include
* memoization support
* migration features for singletons
* other potential requirements

These will be added on later diffs.

Reviewed By: paulbiss

Differential Revision: D21841660

fbshipit-source-id: e4b19a7e1fc52b6387a55cee16d66984304bf37c