Track set of profile TransIDs and reduce TargetProfile counters across them, take 2
Summary:
Due to guard relaxation, a region block in an optimized translation can actually correspond to multiple profile translations that were merged together. When that happened, the block and instructions in the optimized code was only associated with one of those profile translations, and the TargetProfiles were only reduced for that one translation.

This diff keeps track of all the profile translations that are merged together by guard relaxation. The single TransID that was tracked in TransContext, IRGS and BCMarker is replaced by a set of TransIDs. For optimized translations, the TargetProfiles can now have a list of RDS keys and links, which are all reduced together.

Note: this is the same change from D21386699, except that it uses the new destructor support in the arena allocator to free the BCMarker::m_profTransIDs.

Reviewed By: paulbiss

Differential Revision: D21946812

fbshipit-source-id: 8cdb98bea30eeaf5302459942bbbe901246426ea