Remove the sexp dumper of the FFP CST
Summary:
The sexp dumper is a serialisation format that isn't consumed by any tools. It's also tricky to read, and it's not always clear what belongs in it.

The corresponding sexp dumper of the AAST was removed in D17810433.

It's used to debug hackfmt, but following a discussion with jakebailey, we think it would be simpler if we removed it.

Reviewed By: zilberstein

Differential Revision: D21651181

fbshipit-source-id: 9231ec1f6414e733d65a467d7150c4d50a710c6f