Delete TypeHintNotices flag
Summary:
This flag is latched to true, post specialization. (Why true? The HHBBC and JIT analysis of whether typehints may throw is identical for notices and errors, so the simplest way to get it right was to make specialization an extra behavior on top of notices. Setting specialization to true forces notices to true as well.)

Delete the flag's usages. I chose to keep the actual flag in runtime-option for now because it's set in some ~200 tests, and deleting it would cause those tests to fail. All the tests set it to true, so the new code "respects" the test setting.

Reviewed By: alexeyt

Differential Revision: D22111209

fbshipit-source-id: 37d90005a3c91752636cde18249a38e2bf31f1c9